Add notification emailer service
Add checks that when a user likes a movie, the movie submitter gets a
notification email with the appropriate message.

To make the test pass, add an ActionMailer to send the email based on a
template.

TODO: Make the 'hate' version, but that's basically the same.
TODO: Change the user model to include email information